requirements.txt
suds-py3==1.3.2.0
